% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EM_blood}
\alias{EM_blood}
\title{EM algorithm for estimating frequency of A, B, and O blood alleles}
\usage{
EM_blood(A, B, AB, O, tol = 1e-06, verbose = FALSE)
}
\arguments{
\item{A}{number of people with blood type A}

\item{B}{number of people with blood type B}

\item{AB}{number of people with blood type AB}

\item{O}{number of people with blood type O}

\item{tol}{tolerance level governing when to stop the iterations. Here we use the large absolute error for any parameter and compare it to \code{tol}. Iterations stop when the error is less than \code{tol}. Defaults to \code{1e-6}.}

\item{verbose}{Logical; if \code{TRUE} then function prints iterative feedback to the console, and if \code{FALSE} there is no printing. Defaults to \code{FALSE}.}
}
\value{
A data.frame with three entries in one row, giving the estimated allele frequencies of A, B, and O respectively.
}
\description{
Given the phenotypic frequencies for the four blood types (A, B, AB, O), uses the EM algorithm to obtain maximum likelihood estimates for the frequency of the three blood alleles (A, B, O) in a population.
}
\details{
The genotypic frequencies are usually not known for an entire population, because people with blood type A can have a genotype of A/A or A/O, and people with blood type B can have a genotype of B/B or B/O. For blood types AB and O, there is a one-to-one mapping to the genotypes A/B and O/O however. Hence, in the EM setting, the six possible genotypes are viewed as latent variables. Using traditional MLE, closed form solutions cannot obtained.

We assume that the population is in Hardy-Weinberg Equilibrium. That is, for genotype of form X/X, P(X/X) = P(X)^2 and for genotype of form X/Y, P(X/Y) = 2 * P(X) * P(Y).

An initial value of (1/3, 1/3, 1/3) is reasonable as it corresponds to the situation of equal allele frequency. In the function definition, the iterations are updated in the calls to \code{pA_new}, \code{pB_new}, \code{pO_new}. The equations can be verified by going through the Expectation-Maximization algorithm.
}
\examples{
# Population of 100, with equal phenotypic frequencies
A <- 25; B <- 25; AB <- 25; O <- 25
EM_blood(A, B, AB, O)

# Population of 1000, with different phenotypic frequencies
A <- 500; B <- 200; AB <- 50; O <- 250
EM_blood(A, B, AB, O)
}

