% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naim.R
\docType{package}
\name{naim}
\alias{naim}
\alias{naim-package}
\title{naim}
\description{
Numerical analysis iterative methods for maximum likelihood estimates.
}
\details{
Provides alternative procedures for computing maximum likelihood estimates when closed form solutions are unattainable. Current version contains specific implementations of the Newton-Raphson method and the EM algorithm.
}

